solution/api/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
PyJWT==1.7.1
pytest==4.1.1
six==1.12.0
tornado==5.1.1
